Update docs, remove mentions of nw, include package option for nw
